requirements.txt
Flask
python-dotenv
requests
pytest
pytest-mock
WeasyPrint
markdown-it-py
google-api-python-client

flask-sqlalchemy
pgvector
flask-migrate
